feat: Do not always fail when decoding fails
json:api has a bunch of situations where empty answers without
a json to be decoded are perfectly valid. With the advent of
tests for the requesting actions, it became necessary to start
adding code for these valid dataless responses. Additionally,
only a handful of status codes may have data.