Add Actions tests (#244)
* refactor: Move tests to correct nesting level

* style: fix typo

* fix: Repair JsonApiRouter test

* feat: unskip skipped tests

* docs: minor

* docs: use fenced code block

* feat: Fetch Vuex.Store from ResourcefulApi

* fix: Don't use performance api without window

* feat: Make router.addRoute() fluent

* feat: Create a usable api instance for testing

* refactor: Don't pass store to ModuleBuilder twice

* feat: Add basic getAction test

* refactor: Use for-of insteaf of for-in + hasOwn

* docs: Fix docblock

* fix: Resource reading is also allowed with list

* fix: Use new function name

* feat: Apply parameters on method to create valid promise

* refactor: Inline method proxy into resource proxy

* docs: Improve docblock

* style: minor

* feat: Add listAction test

* feat: Reset fetch mock between tests

* feat: Reduce test error-surface

* docs: Default to always show sidebar

* docs: Finally add the necessary list() call to the example

* feat: Do not always fail when decoding fails

json:api has a bunch of situations where empty answers without
a json to be decoded are perfectly valid. With the advent of
tests for the requesting actions, it became necessary to start
adding code for these valid dataless responses. Additionally,
only a handful of status codes may have data.

* feat: basic tests for delete action

* docs: Rewrite and improve upon dev docs

* refactor: inline state objects

It was a stupid past!eFrane idea to split this code base up into so
many files. There are quite a few of there refactorings to come I fear.

* feat: Add ActionError

* feat: Actually test state modification

* feat: Align delete arguments with other actions

* fix: Adjust mocked resource path

* feat: Test store state changes on delete

* feat: Test delete with 200 OK status from server

* feat: Test delete with 202 Accepted

The expected behaviour of accepted on the server side
is that the item will be deleted. In the optimistic
spirit of eventual consistency, the client therefore
treats accepted just like ok and removes the item.

* feat: Add fallback for scalar id delete calls

* refactor: Fold missing module registration into response processing

* fix: use correct type hint

* refactor: Inline validateResourceObject check

* refactor: Only check for basic resource structure

As resource objects with and without client-generated
ids have to be considered for the createAction it does
not make sense to do that check during the basic
structure validation.

* feat: Handle and test the default case for creates

* feat: Handle allowed status codes on create