feat: add request data callbacks
This removes the non-conforming type case wrangling
from the library. Uses where this is needed may still
add it in later as a request data processing callback.

The request callbacks receive all request info in reverse
doAction order, meaning callback(data, parameters, url, method).
url is only the path given to one of the method helpers, not yet
the fully build url if a base url is set on the api object.